EDU-4799 news and merchant article integration

* EDU-4799 news and merchant article integration

* EDU-4799 news and merchant article integration

* EDU-4799 news/article for exam and institute details page

Approved-by: Karan Verma

+++ digital-education-explore/src/main/java/com/paytm/digital/education/explore/request/dto/articles/MerchantArticleRequest.java
@@ -70,7 +70,7 @@ public class MerchantArticleRequest {
     @Min(1)
     @JsonProperty("exam_id")
     @ApiModelProperty(value = "exam id ")
-    private Integer examId;
+    private Long examId;
 
     @NotBlank
     @JsonProperty("stream")
